[web-admin] fix bug change equipment status - change contractor status - Validate staff - number of feedback contractor

+++ webadmin/src/main/java/com/ccp/webadmin/controllers/FeedbackController.java
@@ -62,7 +62,7 @@ public class FeedbackController {
         }
         model.addAttribute("feedbackTypes", feedbackTypeService.findAll());
         feedbackService.save(feedbackEntity);
-        long id = feedbackEntity.getId();
+        Integer id = feedbackEntity.getId();
         return "redirect:detail/" +  id;
     }
